+++ lib/falconmetrics_flutter.dart
@@ -1,3 +1,5 @@
 export 'src/events.dart';
 export 'src/falconmetrics_flutter.dart';
 export 'src/model/tracking_authorization_status.dart';
+export 'src/model/tracking_options.dart';
+export 'src/model/user_data.dart';