+++ lib/src/platform/falconmetrics_flutter_method_channel.dart
@@ -1,5 +1,9 @@
 import 'package:falconmetrics_flutter/src/events.dart';
+import 'package:falconmetrics_flutter/src/generated/userdata.pb.dart'
+    as pb_userdata;
 import 'package:falconmetrics_flutter/src/model/tracking_authorization_status.dart';
+import 'package:falconmetrics_flutter/src/model/tracking_options.dart';
+import 'package:falconmetrics_flutter/src/model/user_data.dart';
 import 'package:falconmetrics_flutter/src/platform/event_proto_converter.dart';
 import 'package:falconmetrics_flutter/src/platform/falconmetrics_flutter_platform_interface.dart';
 import 'package:flutter/foundation.dart';
@@ -14,10 +18,15 @@ class MethodChannelFalconmetricsFlutter extends FalconmetricsFlutterPlatform {
   );
 
   @override
-  Future<void> init({required String apiKey, String? fbAppId}) async {
+  Future<void> init({
+    required String apiKey,
+    TrackingOptions trackingOptions = const TrackingOptions(),
+    String? fbAppId,
+  }) async {
     await methodChannel.invokeMethod<void>('init', {
       'apiKey': apiKey,
       'fbAppId': fbAppId,
+      'ipAddressTracking': trackingOptions.ipAddressTracking?.name,
     });
   }
 
@@ -29,12 +38,27 @@ class MethodChannelFalconmetricsFlutter extends FalconmetricsFlutterPlatform {
   }
 
   @override
-  Future<void> trackEvent({required TrackingEvent event}) async {
+  Future<void> trackEvent({
+    required TrackingEvent event,
+    UserData? userData,
+  }) async {
     final protoEvent = EventProtoConverter().convert(event);
-    await methodChannel.invokeMethod<void>(
-      'trackEvent',
-      protoEvent.writeToBuffer(),
+    final userDataProto = pb_userdata.UserData(
+      email: userData?.email,
+      phoneNumber: userData?.phoneNumber,
+      firstName: userData?.firstName,
+      lastName: userData?.lastName,
+      dateOfBirth: userData?.dateOfBirth,
+      city: userData?.city,
+      state: userData?.state,
+      postalCode: userData?.postalCode,
+      country: userData?.country,
     );
+
+    await methodChannel.invokeMethod<void>('trackEvent', {
+      'event': protoEvent.writeToBuffer(),
+      'userData': userDataProto.writeToBuffer(),
+    });
   }
 
   @override
@@ -80,6 +104,15 @@ class MethodChannelFalconmetricsFlutter extends FalconmetricsFlutterPlatform {
         'getTrackingAuthorizationStatus is only supported on iOS',
       );
     }
+  }
+
+  @override
+  Future<void> updateTrackingOptions({
+    required TrackingOptions trackingOptions,
+  }) async {
+    await methodChannel.invokeMethod<void>('updateTrackingOptions', {
+      'ipAddressTracking': trackingOptions.ipAddressTracking?.name,
+    });
   }
 }