+++ test/platform/falconmetrics_flutter_method_channel_test.dart
@@ -1,4 +1,8 @@
 import 'package:falconmetrics_flutter/src/events.dart';
+import 'package:falconmetrics_flutter/src/generated/userdata.pb.dart'
+    as pb_userdata;
+import 'package:falconmetrics_flutter/src/model/tracking_options.dart';
+import 'package:falconmetrics_flutter/src/model/user_data.dart';
 import 'package:falconmetrics_flutter/src/platform/event_proto_converter.dart';
 import 'package:falconmetrics_flutter/src/platform/falconmetrics_flutter_method_channel.dart';
 import 'package:flutter/services.dart';
@@ -31,11 +35,21 @@ void main() {
 
   test('It calls init with the correct arguments', () async {
     const String testApiKey = 'test_api_key_123';
-    await platform.init(apiKey: testApiKey, fbAppId: '456');
+    await platform.init(
+      apiKey: testApiKey,
+      fbAppId: '456',
+      trackingOptions: TrackingOptions(
+        ipAddressTracking: IpAddressTracking.anonymised,
+      ),
+    );
 
     expect(log, hasLength(1));
     expect(log.first.method, 'init');
-    expect(log.first.arguments, {'apiKey': testApiKey, 'fbAppId': '456'});
+    expect(log.first.arguments, {
+      'apiKey': testApiKey,
+      'fbAppId': '456',
+      'ipAddressTracking': 'anonymised',
+    });
   });
 
   test('It calls setDebugLoggingEnabled with the correct arguments', () async {
@@ -44,6 +58,17 @@ void main() {
     expect(log, hasLength(1));
     expect(log.first.method, 'setDebugLoggingEnabled');
     expect(log.first.arguments, {'enabled': true});
+  });
+  test('It calls updateTrackingOptions with the correct arguments', () async {
+    await platform.updateTrackingOptions(
+      trackingOptions: TrackingOptions(
+        ipAddressTracking: IpAddressTracking.anonymised,
+      ),
+    );
+
+    expect(log, hasLength(1));
+    expect(log.first.method, 'updateTrackingOptions');
+    expect(log.first.arguments, {'ipAddressTracking': 'anonymised'});
   });
 
   test('It calls trackEvent with the correct arguments', () async {
@@ -54,12 +79,20 @@ void main() {
       currency: 'USD',
     );
 
+    final userData = pb_userdata.UserData(email: 'foo@bar.com');
+
     final eventProto = EventProtoConverter().convert(event);
-    await platform.trackEvent(event: event);
+    await platform.trackEvent(
+      event: event,
+      userData: UserData(email: 'foo@bar.com'),
+    );
 
     expect(log, hasLength(1));
     expect(log.first.method, 'trackEvent');
-    expect(log.first.arguments, eventProto.writeToBuffer());
+    expect(log.first.arguments, {
+      'event': eventProto.writeToBuffer(),
+      'userData': userData.writeToBuffer(),
+    });
   });
 
   test('It calls setTrackingEnabled with the correct arguments', () async {
@@ -67,7 +100,7 @@ void main() {
 
     expect(log, hasLength(1));
     expect(log.first.method, 'setTrackingEnabled');
-    expect(log.first.arguments, true);
+    expect(log.first.arguments, {'trackingEnabled': true});
   });
 
   test('It calls isTrackingEnabled with the correct arguments', () async {