+++ test/platform/falconmetrics_flutter_test.dart
@@ -1,5 +1,6 @@
 import 'package:falconmetrics_flutter/src/events.dart';
 import 'package:falconmetrics_flutter/src/falconmetrics_flutter.dart';
+import 'package:falconmetrics_flutter/src/model/tracking_options.dart';
 import 'package:falconmetrics_flutter/src/platform/falconmetrics_flutter_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
@@ -19,6 +20,7 @@ void main() {
 
   setUpAll(() {
     registerFallbackValue(CompleteRegistrationEvent());
+    registerFallbackValue(TrackingOptions());
   });
 
   setUp(() {
@@ -28,12 +30,19 @@ void main() {
       () => FalconmetricsFlutterPlatform.instance.init(
         apiKey: any(named: 'apiKey'),
         fbAppId: any(named: 'fbAppId'),
+        trackingOptions: any(named: 'trackingOptions'),
       ),
     ).thenAnswer((_) => Future.value());
 
     when(
       () => FalconmetricsFlutterPlatform.instance.setTrackingEnabled(
         enabled: any(named: 'enabled'),
+      ),
+    ).thenAnswer((_) => Future.value());
+
+    when(
+      () => FalconmetricsFlutterPlatform.instance.updateTrackingOptions(
+        trackingOptions: any(named: 'trackingOptions'),
       ),
     ).thenAnswer((_) => Future.value());
 
@@ -60,10 +69,24 @@ void main() {
     FalconmetricsFlutterPlatform.instance = initialPlatform;
   });
 
-  test('getPlatformVersion', () async {
-    await sut.init(apiKey: '123', fbAppId: '456');
+  test('initialise', () async {
+    await sut.init(
+      apiKey: '123',
+      fbAppId: '456',
+      trackingOptions: TrackingOptions(
+        ipAddressTracking: IpAddressTracking.full,
+      ),
+    );
 
-    verify(() => sut.init(apiKey: '123', fbAppId: '456')).called(1);
+    verify(
+      () => FalconmetricsFlutterPlatform.instance.init(
+        apiKey: '123',
+        fbAppId: '456',
+        trackingOptions: TrackingOptions(
+          ipAddressTracking: IpAddressTracking.full,
+        ),
+      ),
+    ).called(1);
   });
 
   test('setTrackingEnabled', () async {
@@ -77,10 +100,30 @@ void main() {
     expect(result, false);
   });
 
+  test('updateTrackingOptions', () async {
+    await sut.updateTrackingOptions(
+      trackingOptions: TrackingOptions(
+        ipAddressTracking: IpAddressTracking.full,
+      ),
+    );
+
+    verify(
+      () => FalconmetricsFlutterPlatform.instance.updateTrackingOptions(
+        trackingOptions: TrackingOptions(
+          ipAddressTracking: IpAddressTracking.full,
+        ),
+      ),
+    ).called(1);
+  });
+
   test('setDebugLoggingEnabled', () async {
     await sut.setDebugLoggingEnabled(enabled: true);
 
-    verify(() => sut.setDebugLoggingEnabled(enabled: true)).called(1);
+    verify(
+      () => FalconmetricsFlutterPlatform.instance.setDebugLoggingEnabled(
+        enabled: true,
+      ),
+    ).called(1);
   });
 
   test('Track event', () async {
@@ -94,7 +137,7 @@ void main() {
     );
 
     verify(
-      () => sut.trackEvent(
+      () => FalconmetricsFlutterPlatform.instance.trackEvent(
         event: AddedToCartEvent(
           itemId: '123',
           quantity: 1,